requirements.txt
sqlalchemy
requests
python-dotenv
openai
aiogram
langgraph
langchain
langchain-openai
FlightRadarAPI
